+++ src/samples/sample_wkb_2_b3dm/Program.cs
@@ -14,7 +14,7 @@ namespace sample_wkb_2_b3dm
             var inputfile = @"testfixtures/building.wkb";
             var f = File.OpenRead(inputfile);
             var g = Geometry.Deserialize<WkbSerializer>(f);
-            var translation = new float[] { 539085.1f, 6989220.68f, 52.98f };
+            var translation = new double[] { 539085.1, 6989220.68, 52.98 };
             var glb = GeometryToGlbConvertor.Convert(g, translation);
             var b3dm = GlbToB3dmConvertor.Convert(glb);
             B3dmWriter.WriteB3dm("building.b3dm", b3dm);
